🤖 refactor(site/src/pages): nest deployment user pages
